refactor: move braintrust integration to vendor package
