+++ apps/app/src/env.ts
@@ -2,8 +2,8 @@ import { env as dbEnv } from "@db/app/env";
 import { createEnv } from "@t3-oss/env-nextjs";
 import { vercel } from "@t3-oss/env-nextjs/presets-zod";
 import { clerkEnvBase } from "@vendor/clerk/env";
+import { braintrustEnv } from "@vendor/braintrust/env";
 import { betterstackEnv } from "@vendor/observability/betterstack-env";
-import { braintrustEnv } from "@vendor/observability/braintrust-env";
 import { sentryEnv } from "@vendor/observability/sentry-env";
 import { upstashEnv } from "@vendor/upstash/env";
 import { z } from "zod";
